+++ template/src/MainLayout.js
@@ -6,7 +6,7 @@ import {useSelector} from 'react-redux'
 import Toast from './components/Toast'
 import {getAppStackState, getLoadingIndicator} from './store/selectors'
 import RouteKey from './navigation/RouteKey'
-import {DebugMenu} from './components/debug-menu/DebugMenu'
+import {DebugMenu} from './components/DebugMenu'
 import configs from './constants/configs'
 
 function MainLayout() {
