--- template/src/components/info-menu/styles.js
@@ -1,14 +1,0 @@
-import {StyleSheet} from 'react-native'
-
-export const styles = StyleSheet.create({
-  container: {
-    borderColor: '#151515',
-    paddingBottom: 12,
-    justifyContent: 'space-between',
-  },
-  titleWrapper: {
-    justifyContent: 'space-between',
-    flexDirection: 'row',
-    marginBottom: 8,
-  },
-})
